+++ project/c3d-project/src/lib.rs
@@ -17,7 +17,7 @@ mod template;
 pub use authoring::PrimitiveCreateReport;
 pub use c3d_export_gltf::GltfExportReport;
 pub use c3d_export_gsplat::GsplatExportReport;
-pub use c3d_export_ply::PlyExportReport;
+pub use c3d_export_ply::{PlyExportFormat, PlyExportOptions, PlyExportReport};
 pub use c3d_export_usd::UsdExportReport;
 pub use error::{ProjectError, ProjectResult};
 pub use gsplat::GaussianSplatImportReport;